Save a Sankey diagram to an HTML or PNG file

Save a Sankey diagram created via sankeyNetwork() to an HTML or PNG file. The HTML can include its dependencies in an adjacent directory or can bundle all dependencies into the HTML file (via base64 encoding). PNG files are converted from HTML using the webshot package.

saveNetwork(
  network,
  file,
  selfcontained = TRUE,
  vwidth = network$width,
  vheight = network$height,
  zoom = 5,
  delay = 1
)

Arguments

network

Network to save, i.e. result of sankeyNetwork().

file

file to be saved - either html or png.

selfcontained

Whether to save the HTML as a single self-contained file (with external resources base64 encoded) or a file with external resources placed in an adjacent directory.

vwidth

Viewport width in pixels. Defaults to network$width.

vheight

Viewport height in pixels. Defaults to network$height.

zoom

Zoom factor. 1 means actual size, 2 means 50% zoom, etc.

delay

Time to delay before taking webshot, in seconds.